requirements.txt
fastapi
httpx
pydantic
python-dotenv
stripe
uvicorn
